+++ src/Symfony/Component/DependencyInjection/Loader/YamlFileLoader.php
@@ -129,7 +129,7 @@ class YamlFileLoader extends FileLoader
 
             if (!$this->prepend && !$this->container->hasExtension($namespace)) {
                 $extensionNamespaces = array_filter(array_map(static fn (ExtensionInterface $ext) => $ext->getAlias(), $this->container->getExtensions()));
-                throw new InvalidArgumentException(UndefinedExtensionHandler::getErrorMessage($namespace, $file, $namespace, $extensionNamespaces));
+                throw new InvalidArgumentException(UndefinedExtensionHandler::getErrorMessage($namespace, $file, $namespace, $extensionNamespaces, UndefinedExtensionHandler::getPackages($this->container)));
             }
         }